+++ tedi/components/navigation/horizontal-stepper/horizontal-stepper-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./horizontal-stepper-item.component";